Factor out a `ViewAsUserMenuItem` component
